Feat: 페이지네이션 공통 DTO, Service 생성

+++ src/main/java/com/project/hrbank/service/EmployeeLogService.java
@@ -13,7 +13,7 @@ public interface EmployeeLogService {
 
 	String getLogById(Long id);
 
-	Integer getLogCount();
+	long getLogCount();
 
 	EmployeeLogResponse createLog(EmployeeLogRequest request);
